Add ARIA, and made navigation responsive on mobile
Created a style that allows navigation flow in on mobile phone. When the toggle buttton is click, it makes body overflow to be hidden (no scroll). on media query of maximum height of 450px, the nav height is 100% and overflow is automatic. ARIA attribute (aria=hidden) if true sets the dsplay as none while false, sets visibility to visibile. I used visible so as not to affect the element display CSS.