Add property generation function
Creates a function that gets all the blog field items and values from the JSON.
Then, it call a createBlogHTML function which check the blog.structure.type to know if it's a p, h1, h2, h3, h4, ul, ol or article. from there, it creates the need HTML.

I used a loop to further create HTML for those that has an array of content with children type.

It was fun when it ran immediately with little error.